requirements.txt
anyio==4.6.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.3.2
et-xmlfile==1.1.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
numpy==2.1.1
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
python-telegram-bot==21.6
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
tzdata==2024.2
urllib3==2.2.3
